orchestrator/requirements.txt

fastapi==0.108.0
uvicorn[standard]==0.25.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
pydantic==2.5.3
python-dotenv==1.0.0
loguru==0.7.2
celery[redis]==5.3.6
httpx==0.26.0
pytest==7.4.3
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
